Add formatHTML module
