+++ _scripts/app.js
@@ -12,6 +12,7 @@ var formatJSON = require('./_outputs/format-JSON');
 var formatCSV = require('./_outputs/format-CSV');
 var formatMD = require('./_outputs/format-MD');
 var formatXML = require('./_outputs/format-XML');
+var formatHTML = require('./_outputs/format-HTML');
 
 // Wait for the document to load, then load settings for the user, then run the app.
 $(document).ready(function () {
@@ -576,9 +577,9 @@ function runApp () {
                 successMessage(file, filetype);
             // Markdown
             } else if (ugui.args.outputmd.htmlticked) {
-                var output = formatMD(window.imageStats, results, ugui.args.url.value);
+                var outputMD = formatMD(window.imageStats, results, ugui.args.url.value);
 
-                ugui.helpers.writeToFile(file, output);
+                ugui.helpers.writeToFile(file, outputMD);
 
                 successMessage(file, filetype);
             // XML
@@ -590,100 +591,24 @@ function runApp () {
                 successMessage(file, filetype);
             // HTML
             } else {
-                var returnedErrors = '';
-                var returnedWarnings = '';
-                var returnedNotices = '';
-                var panelColor = '';
-                for (i = 0; i < results.length; i++) {
-                    var resultsType = results[i].type;
-                    if (resultsType == 'warning') {
-                        panelColor = 'warning';
-                    } else if (resultsType == 'error') {
-                        panelColor = 'danger';
-                    } else if (resultsType == 'notice') {
-                        panelColor = 'primary';
-                    }
-
-                    var theContext = results[i].context;
-                    theContext = theContext.split('<').join('&lt;');
-                    var theMessage = results[i].message;
-                    theMessage = theMessage.replace('. Recommendation: ', '. <strong>Recommendation:</strong> ');
-                    var entry =
-                      '<div class="panel panel-' + panelColor + '">\n' +
-                        '<div class="panel-heading">' + results[i].code + '</div>\n' +
-                        '<div class="panel-body">\n' +
-                          '<strong class="text-capitalize">' + results[i].type + ':</strong> ' + theMessage + '<br /><br />\n' +
-                          '<pre><code>' + theContext + '</code></pre>\n' +
-                        '</div>\n' +
-                        '<div class="panel-footer text-sm"><h4><small>' + results[i].selector + '</small></h4></div>\n' +
-                      '</div>\n';
-
-                    if (resultsType == 'error') {
-                        returnedErrors = returnedErrors + entry;
-                    } else if (resultsType == 'warning') {
-                        returnedWarnings = returnedWarnings + entry;
-                    } else if (resultsType == 'notice') {
-                        returnedNotices = returnedNotices + entry;
+                var enabledButtons = [];
+                var buttonElements = document.getElementById('button-badges').children;
+                for (var keys in buttonElements) {
+                    if (typeof(buttonElements[keys]) === 'object') {
+                        var val = buttonElements[keys].outerHTML;
+                        // If not disabled
+                        if (val.indexOf('disabled') === -1) {
+                            enabledButtons.push(val);
+                        }
                     }
                 }
+                var buttons = enabledButtons.join('');
 
-                $.get('_markup/template.html', function (template) {
-                    var results = returnedErrors + returnedWarnings + returnedNotices;
-                    var buttons = '';
-                    $('#button-badges button:not(".disabled")').each(function () {
-                        buttons = buttons + $(this).prop('outerHTML') + '\n';
-                    });
+                var outputHTML = formatHTML(window.imageStats, results, url, buttons);
 
-                    var imgAlts = '';
+                ugui.helpers.writeToFile(file, outputHTML);
 
-                    // Ensure that the imageStats Object is not empty
-                    if (!$.isEmptyObject(window.imageStats)) {
-                        var totalImages = window.imageStats.totalImages;
-                        var descriptive = window.imageStats.descriptive;
-                        var descriptivePercent = window.imageStats.descriptivePercent;
-                        var under100Char = window.imageStats.under100Char;
-                        var under100CharPercent = window.imageStats.under100CharPercent;
-                        var under100KB = window.imageStats.under100KB;
-                        var under100KBPercent = window.imageStats.under100KBPercent;
-                        var imagesLoaded = window.imageStats.imagesLoaded;
-                        var imagesLoadedPercent = window.imageStats.imagesLoadedPercent;
-                        var totalFileSizeInKB = window.imageStats.totalFileSizeInKB;
-                        var descriptiveStyle = 'success glyphicon-ok';
-                        var under100CharStyle = 'success glyphicon-ok';
-                        var under100KBStyle = 'success glyphicon-ok';
-                        var imagesLoadedStyle = 'success glyphicon-ok';
-                        if (descriptivePercent  < 100) { descriptiveStyle  = 'danger glyphicon-remove'; }
-                        if (under100CharPercent < 100) { under100CharStyle = 'danger glyphicon-remove'; }
-                        if (under100KBPercent   < 100) { under100KBStyle   = 'danger glyphicon-remove'; }
-                        if (imagesLoadedPercent < 100) { imagesLoadedStyle = 'danger glyphicon-remove'; }
-
-                        imgAlts =
-                          '<div class="row">' +
-                            '<div class="panel panel-primary">' +
-                              '<div class="panel-heading">Image Accessibility</div>' +
-                              '<div class="panel-body">' +
-                                '<p><i class="glyphicon text-' + descriptiveStyle  + '"></i> <strong>' + descriptivePercent  + '%</strong> of images on the page had descriptive ALT text. <strong>(' + descriptive + '/' + totalImages + ')</strong></p>' +
-                                '<p><i class="glyphicon text-' + under100CharStyle + '"></i> <strong>' + under100CharPercent + '%</strong> of ALTs were under 100 characters. <strong>(' + under100Char + '/' + totalImages + ')</strong></p>' +
-                                '<p><i class="glyphicon text-' + under100KBStyle   + '"></i> <strong>' + under100KBPercent   + '%</strong> of images were under 100KB in size. <strong>(' + under100KB + '/' + totalImages + ')</strong></p>' +
-                                '<p><i class="glyphicon text-' + imagesLoadedStyle + '"></i> <strong>' + imagesLoadedPercent + '%</strong> of images loaded with a total image payload of <strong>' + totalFileSizeInKB + 'KB (' + imagesLoaded + '/' + totalImages + ')</strong></p>' +
-                              '</div>' +
-                            '</div>' +
-                          '</div>';
-                    }
-
-                    var content =
-                        '    <div class="row">\n' +
-                        '      <span id="buttons">' + buttons + '</span>\n' +
-                        '      <h1>' + url + '</h1>\n' +
-                        '    </div>\n' +
-                             imgAlts + '\n' +
-                        '    <div class="row">' + results + '</div>\n';
-                    var output = template.replace('<!-- Content goes here -->', content);
-
-                    ugui.helpers.writeToFile(file, output);
-
-                    successMessage(file, filetype);
-                });
+                successMessage(file, filetype);
             }
         });
     }
